Fix test.sh only loading Breeze defaults
Initially, test.sh functionality was verified on the basis
that it loaded the default Breeze theme's SDDM and lockscreen --
as this was being worked on in Blackbriar KDE's directories.

However, no changes to the QML were registered during actual work.

Calling the functions outlined in testing_commands, via a shell
script, on a theme obviously different to Breeze on my local machine
```bash
/usr/lib/kscreenlocker_greet --testing --theme some-other-obvious-theme
```
worked just fine. However, in this specific script, the calls failed
and only loaded defaults.

Oddly, it was found that piping the last line to bash, rather than
calling it directly with the dollar sign, made the SDDM theme succeed.

However, the lockscreen theme had another bug: it could load other themes,
but not the one in this current directory. By testing relative and
absolute paths, it was found that kscreenlocker_greet needed an
absolute path to work correctly.

Summarily, the testing script no longer falsely shows default Breeze
theming under the guise of successfully working.