fix(lock): eye open/close always appears
The eye icon to view and hide the password either appeared successfully
in testing but not in installation, or vice versa, leaving a placeholder
icon in its place. This is because a string was being used to access the
icon's file.

To fix this, a Qt.resolvedUrl was used to point to the pertinent SVGs.