feat(lock/sddm): minimize CustomPasswordField deps
Minimizing imports in blackbriar-components will contribute to ease
in making the lockscreen and sddm screen look the same.

Note that this removes the translucent blur feature in
CustomPasswordField.qml itself. It may still be able to be added on
as an attribute *after* the object gets instantiated, but this will not
be a primary aim of the theme.